+++ wpull/app_test.py
@@ -1,6 +1,7 @@
 # encoding=utf-8
 import glob
 from http import cookiejar
+import io
 import re
 import gzip
 import hashlib
@@ -253,6 +254,27 @@ class TestApp(GoodAppTestCase, TempDirMixin):
 
         self.assertEqual(0, exit_code)
         self.assertEqual(builder.factory['Statistics'].files, 2)
+
+    @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
+    def test_app_input_file_arg_stdin(self):
+        arg_parser = AppArgumentParser(real_exit=False)
+
+        real_stdin = sys.stdin
+        fake_stdin = io.StringIO(self.get_url('/') + '\n')
+
+        try:
+            sys.stdin = fake_stdin
+            args = arg_parser.parse_args([
+                '--input-file', '-'
+            ])
+            builder = Builder(args, unit_test=True)
+            app = builder.build()
+            exit_code = yield From(app.run())
+        finally:
+            sys.stdin = real_stdin
+
+        self.assertEqual(0, exit_code)
+        self.assertEqual(builder.factory['Statistics'].files, 1)
 
     @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
     def test_app_args_warc_size(self):
@@ -1214,6 +1236,58 @@ class TestApp(GoodAppTestCase, TempDirMixin):
         # FIXME: version 2015.01.25 doesn't have thumbnail?
         # thumbnails = tuple(glob.glob('*.jpg'))
         # self.assertTrue(thumbnails)
+
+    @unittest.skip('not a good idea to test continuously on external servers')
+    @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
+    def test_propagate_ipv4_only_and_no_cert_check_to_youtube_dl(self):
+        arg_parser = AppArgumentParser()
+        args = arg_parser.parse_args([
+            'https://www.youtube.com/watch?v=tPEE9ZwTmy0',
+            '--warc-file', 'test',
+            '--debug',  # to capture youtube-dl arguments in the log
+            '--no-warc-compression',
+            '--youtube-dl',
+            '--inet4-only',
+            '--no-check-certificate',
+            '--output-file', 'test.log'
+            ])
+        builder = Builder(args, unit_test=True)
+
+        app = builder.build()
+        exit_code = yield From(app.run())
+
+        self.assertEqual(0, exit_code)
+
+        with open('test.log', 'rb') as test_log:
+            data = test_log.read()
+
+            self.assertTrue(re.search(b'Starting process \[\'youtube-dl.*--force-ipv4', data))
+            self.assertTrue(re.search(b'Starting process \[\'youtube-dl.*--no-check-certificate', data))
+
+    @unittest.skip('not a good idea to test continuously on external servers')
+    @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
+    def test_youtube_dl_defaults_have_neither_ipv4_only_nor_no_cert_check(self):
+        arg_parser = AppArgumentParser()
+        args = arg_parser.parse_args([
+            'https://www.youtube.com/watch?v=tPEE9ZwTmy0',
+            '--warc-file', 'test',
+            '--debug',
+            '--no-warc-compression',
+            '--youtube-dl',
+            '--output-file', 'test.log'
+            ])
+        builder = Builder(args, unit_test=True)
+
+        app = builder.build()
+        exit_code = yield From(app.run())
+
+        self.assertEqual(0, exit_code)
+
+        with open('test.log', 'rb') as test_log:
+            data = test_log.read()
+
+            self.assertFalse(re.search(b'Starting process \[\'youtube-dl.*--force-ipv4', data))
+            self.assertFalse(re.search(b'Starting process \[\'youtube-dl.*--no-check-certificate', data))
 
     @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
     def test_no_cache_arg(self):