+++ wpull/builder.py
@@ -542,8 +542,11 @@ class Builder(object):
 
     def _read_input_file_as_lines(self):
         '''Read lines from input file and return them.'''
-        input_file = codecs.getreader(
-            self._args.local_encoding or 'utf-8')(self._args.input_file)
+        if self._args.input_file == sys.stdin:
+            input_file = self._args.input_file
+        else:
+            input_file = codecs.getreader(
+                self._args.local_encoding or 'utf-8')(self._args.input_file)
 
         urls = (line.strip() for line in input_file if line.strip())
 
@@ -1359,6 +1362,8 @@ class Builder(object):
             root_path=self._args.directory_prefix,
             user_agent=self._args.user_agent or self.default_user_agent,
             warc_recorder=self._factory.get('WARCRecorder'),
+            inet_family=self._args.inet_family,
+            check_certificate=self._args.check_certificate
         )
 
         return coprocessor